Track tabs with gistdeck running to avoid doubling-up on the script.
Requires a background script that listens for requests. Since the
'sender' is popup.html, the tab must be retrieved with tabs.query().
The active tab in the current window is retrieved (all others
ignored if more than one returned) and its ID becomes a property
of an object with its value a boolean to whether it is running or
not.

The execution of gistdeck.js is setup to run if gistdeck.js returns
a truthy value, since tabs.executeScript() sets the first parameter
of the callback function to the return value of the script called.

This result can be determined in gistdeck.js by checking the length
of the slides set. As a result, this extension will always be
packaging its own maintained gistdeck script, which ain't so bad =)